+++ backend/database/migrations/2023_08_21_001926_create_promotions_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->dateTime('starts_at');
             $table->dateTime('end_at');
-            $table->unsignedInteger('establishment_id')->nullable();
+            $table->unsignedBigInteger('establishment_id')->nullable();
             $table->foreign('establishment_id')->references('id')->on('establishments');
             $table->timestamps();
         });
